Refactored Header with settings
