+++ src/ducks/Studio/Chart/ChartFullscreenBtn.module.scss
@@ -16,3 +16,7 @@
 .content {
   padding: 15px;
 }
+
+.settings {
+  justify-content: flex-end;
+}
