+++ src/ducks/Studio/Header/Settings.module.scss
@@ -1,0 +1,8 @@
+.wrapper {
+  display: flex;
+  align-items: center;
+}
+
+.calendar {
+  margin: 0 16px;
+}
